Fix save-after-load bug and refactor components into directories
