+++ src/components/editor/EditorTabs.vue
