+++ src/components/molecules/FeatureDiscovery.vue
