+++ src/components/molecules/Info.vue
