+++ src/components/molecules/Note.vue
